update return types and added documentation for GamePhases

+++ src/model/gamePhases/GamePhase.java
@@ -3,9 +3,24 @@ package model.gamePhases;
 import model.PhaseTag;
 
 public abstract class GamePhase {
+	/**
+	 * @return The PhaseTag representing this game phase.
+	 */
 	public abstract PhaseTag getPhaseTag();
-	public abstract int getDuration();
-	public abstract int getUpdatePeriod();
+	
+	/**
+	 * @return The duration in millis of this game phase.
+	 */
+	public abstract long getDuration();
+	
+	/**
+	 * @return The period in millis to periodically update the player scores.
+	 */
+	public abstract long getUpdatePeriod();
+	
+	/**
+	 * @return The maximum achievable score of this game phase.
+	 */
 	public abstract int getMaxScore();
 	
 	/** 
@@ -18,6 +33,10 @@ public abstract class GamePhase {
 	          * (float) getMaxScore();
 	}
 	
+	/**
+	 * @param tag the PhaseTag of the desired GamePhase.
+	 * @return The GamePhase representative of the TAG.
+	 */
 	public static GamePhase makeGamePhase(PhaseTag tag) {
 		switch (tag) {
 			case CONCENTRATE:
